# Conflicts:
#	.idea/libraries/Maven__io_jsonwebtoken_jjwt_0_6_0.xml
#	.idea/libraries/Maven__javax_xml_bind_jaxb_api_2_3_0.xml
#	server/src/main/java/com/server/server/api/ShipResource.java
#	server/src/main/java/com/server/server/controllers/LoginController.java
#	server/src/main/java/com/server/server/entity/Ship.java
#	server/src/main/java/com/server/server/repository/ShipRepository.java
#	server/src/main/java/com/server/server/security/WebSecurityConfig.java
#	server/src/main/java/com/server/server/security/jwt/JwtUtils.java
#	server/src/main/java/com/server/server/service/ShipService.java
#	server/src/main/java/com/server/server/service/ShipServiceImpl.java
#	server/src/main/resources/application.properties

+++ server/src/main/java/com/server/server/payload/response/JwtResponse.java
@@ -5,12 +5,12 @@ import java.util.List;
 public class JwtResponse {
     private String token;
     private String type = "Bearer";
-    private Integer id;
+    private Long id;
     private String username;
     private String email;
     private List<String> roles;
 
-    public JwtResponse(String accessToken, Integer id, String username, String email, List<String> roles) {
+    public JwtResponse(String accessToken, Long id, String username, String email, List<String> roles) {
         this.token = accessToken;
         this.id = id;
         this.username = username;
@@ -34,11 +34,11 @@ public class JwtResponse {
         this.type = tokenType;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
