+++ server/src/main/java/com/server/server/repository/RoleRepository.java
@@ -1,13 +1,13 @@
 package com.server.server.repository;
 
-import com.server.server.entity.ERole;
-import com.server.server.entity.Role;
+import com.server.server.model.ERole;
+import com.server.server.model.Role;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
-    Optional<Role> findByName(ERole name);
+public interface RoleRepository extends JpaRepository<Role, Long> {
+   Optional<Role> findByName(ERole name);
 }
