+++ server/src/main/java/com/server/server/repository/UserRepository.java
@@ -1,18 +1,16 @@
 package com.server.server.repository;
 
-import com.server.server.entity.User;
+import com.server.server.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
-    Optional<User> findByUsername(String username);
+public interface UserRepository extends JpaRepository<User, Long> {
+   Optional<User> findByUsername(String username);
 
-    Boolean existsByUsername(String username);
+   Boolean existsByUsername(String username);
 
-    Boolean existsByEmail(String email);
-
-
+   Boolean existsByEmail(String email);
 }
